refactor(neuron-ui): rename context files by Context Name Convention
