+++ packages/neuron-ui/package.json
@@ -50,6 +50,7 @@
     "typescript": "3.2.2"
   },
   "devDependencies": {
+    "@nervosnetwork/ckb-types": "^0.0.1-alpha.1",
     "@types/enzyme": "^3.1.16",
     "@types/enzyme-adapter-react-16": "^1.0.3",
     "@types/qrcode.react": "0.8.2",
