+++ packages/neuron-ui/tsconfig.json
@@ -30,7 +30,7 @@
     "jsx": "preserve",
     "typeRoots": [
       "./types",
-      "node_modules/@types"
+      "node_modules/@types",
     ]
   },
   "include": [
